Structure files. Fixed a lot of small issues.
